+++ src/lib.rs
@@ -16,271 +16,25 @@ macro_rules! c_str_to_str {
     }
 }
 
-// LLVM Wrappers
+mod context;
+mod types;
+mod builder;
+mod module;
 
-pub struct Context {
-    pub context: LLVMContextRef
-}
+pub use context::*;
+pub use types::*;
+pub use builder::*;
+pub use module::*;
 
-impl Context {
-    pub fn new() -> Self {
-        let context = unsafe {
-            llvm::LLVMContextCreate()
-        };
-        Context { context: context }
-    }
+pub fn function_get_param(func: LLVMValueRef, index: u32) -> Option<LLVMValueRef> {
+    let p = unsafe {
+        llvm::LLVMGetParam(func, index)
+    };
 
-    pub fn create_builder(&self) -> Builder {
-        let builder = unsafe {
-            llvm::LLVMCreateBuilderInContext(self.context)
-        };
-        Builder { builder: builder }
-    }
-
-    pub fn module_create_with_name(&self, name: &str) -> Module {
-        let c_name = CString::new(name).unwrap();
-        let module = unsafe {
-            llvm::LLVMModuleCreateWithNameInContext(c_name.as_ptr(), self.context)
-        };
-        Module { module: module }
-    }
-
-    pub fn void_type(&self) -> LLVMTypeRef {
-        unsafe {
-            llvm::LLVMVoidTypeInContext(self.context)
-        }
-    }
-
-    pub fn int64_type(&self) -> LLVMTypeRef {
-        unsafe {
-            llvm::LLVMInt64TypeInContext(self.context)
-        }
-    }
-
-    pub fn int32_type(&self) -> LLVMTypeRef {
-        unsafe {
-            llvm::LLVMInt32TypeInContext(self.context)
-        }
-    }
-
-    pub fn int16_type(&self) -> LLVMTypeRef {
-        unsafe {
-            llvm::LLVMInt16TypeInContext(self.context)
-        }
-    }
-
-    pub fn int8_type(&self) -> LLVMTypeRef {
-        unsafe {
-            llvm::LLVMInt8TypeInContext(self.context)
-        }
-    }
-
-    pub fn int1_type(&self) -> LLVMTypeRef {
-        unsafe {
-            llvm::LLVMInt1TypeInContext(self.context)
-        }
-    }
-
-    pub fn const_bool(&self, val: bool) -> LLVMValueRef {
-        let ty = self.int1_type();
-        const_int(ty, val as u64, false)
-    }
-
-    pub fn append_basic_block(&self, func: LLVMValueRef, name: &str) -> LLVMBasicBlockRef {
-        let c_name = CString::new(name).unwrap();
-        unsafe {
-            llvm::LLVMAppendBasicBlockInContext(self.context, func, c_name.as_ptr())
-        }
-    }
-}
-
-impl Drop for Context {
-    fn drop(&mut self) {
-        unsafe {
-            llvm::LLVMContextDispose(self.context);
-        }
-    }
-}
-
-// No `Drop` impl is needed as this is disposed of when the associated context is disposed
-pub struct Module {
-    pub module: LLVMModuleRef
-}
-
-impl Module {
-    pub fn dump(&self) {
-        unsafe {
-            llvm::LLVMDumpModule(self.module)
-        }
-    }
-
-    pub fn add_function(&mut self, func_ty: LLVMTypeRef, name: &str) -> LLVMValueRef {
-        let c_name = CString::new(name).unwrap();
-        unsafe {
-            llvm::LLVMAddFunction(self.module, c_name.as_ptr(), func_ty)
-        }
-    }
-
-    pub fn get_named_function(&mut self, name: &str) -> Option<LLVMValueRef> {
-        let c_name = CString::new(name).unwrap();
-        let res = unsafe {
-            llvm::LLVMGetNamedFunction(self.module, c_name.as_ptr())
-        };
-
-        if res.is_null() {
-            None
-        } else {
-            Some(res)
-        }
-    }
-}
-
-pub struct Builder {
-    pub builder: LLVMBuilderRef
-}
-
-impl Builder {
-    pub fn position_at_end(&mut self, basic_block: LLVMBasicBlockRef) {
-        unsafe {
-            llvm::LLVMPositionBuilderAtEnd(self.builder, basic_block);
-        }
-    }
-
-    pub fn build_ret(&mut self, ret_val: LLVMValueRef) -> LLVMValueRef {
-        unsafe {
-            llvm::LLVMBuildRet(self.builder, ret_val)
-        }
-    }
-
-    pub fn build_ret_void(&self) -> LLVMValueRef {
-        unsafe {
-            llvm::LLVMBuildRetVoid(self.builder)
-        }
-    }
-
-    pub fn build_alloca(&mut self, ty: LLVMTypeRef, name: &str) -> LLVMValueRef {
-        let c_name = CString::new(name).unwrap();
-        unsafe {
-            llvm::LLVMBuildAlloca(self.builder, ty, c_name.as_ptr())
-        }
-    }
-
-    pub fn build_store(&mut self, val: LLVMValueRef, ptr: LLVMValueRef) -> LLVMValueRef {
-        unsafe {
-            llvm::LLVMBuildStore(self.builder, val, ptr)
-        }
-    }
-
-    pub fn build_load(&mut self, ptr: LLVMValueRef, name: &str) -> LLVMValueRef {
-        let c_name = CString::new(name).unwrap();
-        unsafe {
-            llvm::LLVMBuildLoad(self.builder, ptr, c_name.as_ptr())
-        }
-    }
-
-    pub fn build_call(&mut self, func: LLVMValueRef, mut args: Vec<LLVMValueRef>,
-                      name: &str) -> LLVMValueRef {
-        let c_name = CString::new(name).unwrap();
-        unsafe {
-            llvm::LLVMBuildCall(self.builder, func, args.as_mut_ptr(), args.len() as u32,
-                                c_name.as_ptr())
-        }
-    }
-
-    pub fn build_add(&mut self, lhs: LLVMValueRef, rhs: LLVMValueRef, name: &str) -> LLVMValueRef {
-        let c_name = CString::new(name).unwrap();
-        unsafe {
-            llvm::LLVMBuildAdd(self.builder, lhs, rhs, c_name.as_ptr())
-        }
-    }
-
-    pub fn build_sub(&mut self, lhs: LLVMValueRef, rhs: LLVMValueRef, name: &str) -> LLVMValueRef {
-        let c_name = CString::new(name).unwrap();
-        unsafe {
-            llvm::LLVMBuildSub(self.builder, lhs, rhs, c_name.as_ptr())
-        }
-    }
-
-    pub fn build_mul(&mut self, lhs: LLVMValueRef, rhs: LLVMValueRef, name: &str) -> LLVMValueRef {
-        let c_name = CString::new(name).unwrap();
-        unsafe {
-            llvm::LLVMBuildMul(self.builder, lhs, rhs, c_name.as_ptr())
-        }
-    }
-
-    pub fn build_sdiv(&mut self, lhs: LLVMValueRef, rhs: LLVMValueRef,
-                      name: &str) -> LLVMValueRef {
-        let c_name = CString::new(name).unwrap();
-        unsafe {
-            llvm::LLVMBuildSDiv(self.builder, lhs, rhs, c_name.as_ptr())
-        }
-    }
-
-    pub fn build_icmp(&mut self, op: LLVMIntPredicate, lhs: LLVMValueRef, rhs: LLVMValueRef,
-                      name: &str) -> LLVMValueRef {
-        let c_name = CString::new(name).unwrap();
-        unsafe {
-            llvm::LLVMBuildICmp(self.builder, op, lhs, rhs, c_name.as_ptr())
-        }
-    }
-
-    pub fn build_global_string(&self, s: &str, name: &str) -> LLVMValueRef {
-        let c_s = CString::new(s).unwrap();
-        let c_name = CString::new(name).unwrap();
-        unsafe {
-            llvm::LLVMBuildGlobalString(self.builder, c_s.as_ptr(), c_name.as_ptr())
-        }
-    }
-
-    pub fn build_in_bounds_gep(&self, ptr: LLVMValueRef, mut indices: Vec<LLVMValueRef>,
-                               name: &str) -> LLVMValueRef {
-        let c_name = CString::new(name).unwrap();
-        unsafe {
-            llvm::LLVMBuildInBoundsGEP(self.builder, ptr, indices.as_mut_ptr(),
-                                       indices.len() as u32, c_name.as_ptr())
-        }
-    }
-
-    pub fn build_cond_br(&self, cond: LLVMValueRef, then: LLVMBasicBlockRef,
-                         else_: LLVMBasicBlockRef) -> LLVMValueRef {
-        unsafe {
-            llvm::LLVMBuildCondBr(self.builder, cond, then, else_)
-        }
-    }
-
-    pub fn build_br(&self, dest: LLVMBasicBlockRef) -> LLVMValueRef {
-        unsafe {
-            llvm::LLVMBuildBr(self.builder, dest)
-        }
-    }
-
-}
-
-impl Drop for Builder {
-    fn drop(&mut self) {
-        unsafe {
-            llvm::LLVMDisposeBuilder(self.builder);
-        }
-    }
-}
-
-pub fn const_int(ty: LLVMTypeRef, val: u64, signed: bool) -> LLVMValueRef {
-    unsafe {
-        llvm::LLVMConstInt(ty, val, signed as i32)
-    }
-}
-
-pub fn function_type(ret_ty: LLVMTypeRef, mut param_types: Vec<LLVMTypeRef>,
-                     is_var_args: bool) -> LLVMTypeRef {
-    unsafe {
-        llvm::LLVMFunctionType(ret_ty, param_types.as_mut_ptr(), param_types.len() as u32,
-                               is_var_args as i32)
-    }
-}
-
-pub fn pointer_type(ty: LLVMTypeRef, address_space: u32) -> LLVMTypeRef {
-    unsafe {
-        llvm::LLVMPointerType(ty, address_space)
+    if p.is_null() {
+        return None;
+    } else  {
+        return Some(p);
     }
 }
